+++ apps/OboeTester/app/src/main/cpp/InterpolatingDelayLine.h
@@ -22,7 +22,7 @@
 #include <sys/types.h>
 
 #include "oboe/Oboe.h"
-#include "FullDuplexStream.h"
+#include "common/FullDuplexStream.h"
 
 /**
  * Monophonic delay line.
